refactor: create util functions to  utils and get_session to dependencies
