+++ app/database.py
@@ -1,12 +1,7 @@
-from sqlmodel import Session, create_engine
+from sqlmodel import create_engine
 
 sqlite_file_name = "database.db"
 sqlite_url = f"sqlite:///{sqlite_file_name}"
 
 connect_args = {"check_same_thread": False}
 engine = create_engine(sqlite_url, echo=True, connect_args=connect_args)
-
-
-def get_session():
-    with Session(engine) as session:
-        yield session
