+++ app/dependencies.py
@@ -1,4 +1,13 @@
+from sqlmodel import Session
+
 from app.config import UPLOAD_DIR
+
+from .database import engine
+
+
+def get_session():
+    with Session(engine) as session:
+        yield session
 
 
 def get_upload_dir():
